tests/wpt/web-platform-tests/performance-timeline/tentative/include-frames-one-same-origin-child-one-same-origin-grandchild.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
</body>
<script>
promise_test(() => {
    return new Promise(resolve => {
        performance.clearResourceTimings()

        // Create child iframe with an embedded frame that is cross-origin with its parent, but same-origin with the current frame.
        const childFrame = document.createElement('iframe')
        childFrame.src = "../resources/parent-frame-with-cross-origin-child.sub.html"
        document.body.appendChild(childFrame)

        // Listen for postMessage() from grandchild frame.
        window.addEventListener("message", () => {
            const entries = performance.getEntries(true)
            const navigationEntries = performance.getEntriesByType("navigation", true)
            const markedEntries = performance.getEntriesByName("entry-name", undefined, true)

            // 4 entries for parent, 3 for child, 2 for grandchild.
            assert_equals(entries.length, 9)

            // 1 entry for parent, 1 for child, 1 for grandchild.
            assert_equals(navigationEntries.length, 3)

            // 1 entry for child, 1 for grandchild.
            assert_equals(markedEntries.length, 2)

            resolve()
        })
    })
}, "GetEntries of a parent Frame with one Same-Origin child and one Same-Origin grandchild")
</script>
